requirements.txt
faker
python-dotenv
openai